Move the order events to the order package
