+++ saleor/checkout/utils.py
@@ -23,8 +23,7 @@ from ..discount.models import NotApplicable, Voucher
 from ..discount.utils import (
     get_products_voucher_discount, get_shipping_voucher_discount,
     get_value_voucher_discount, increase_voucher_usage)
-from ..events import OrderEventsEmails
-from ..events.order import OrderEventManager
+from ..order.events import order_created_event
 from ..order.emails import send_order_confirmation
 from ..order.models import Order
 from ..shipping.models import ShippingMethod
@@ -880,15 +879,11 @@ def create_order(
     # remove checkout after order is created
     checkout.delete()
 
-    # Create the order placed and email confirmation sent events
-    OrderEventManager()\
-        .order_created_event(order, user)\
-        .email_sent_event(
-            order=order, email_type=OrderEventsEmails.ORDER, user=user)\
-        .save()
+    # Create the order placed
+    order_created_event(order=order, user=user)
 
     # Send the order confirmation email
-    send_order_confirmation.delay(order.pk)
+    send_order_confirmation.delay(order.pk, user.pk)
     return order
