--- saleor/events/models/__init__.py
@@ -1,4 +1,0 @@
-from .order import OrderEvent
-
-__all__ = [
-    "OrderEvent"]
