+++ saleor/graphql/order/mutations/orders.py
@@ -3,8 +3,10 @@ from django.core.exceptions import ValidationError
 
 from ....account.models import User
 from ....core.utils.taxes import ZERO_TAXED_MONEY
-from ....events.order import OrderEventManager
 from ....order import models
+from ....order.events import (
+    order_note_added_event, payment_captured_event, payment_failed_event,
+    payment_voided_event)
 from ....order.utils import cancel_order
 from ....payment import CustomPaymentChoices, PaymentError
 from ....payment.utils import (
@@ -75,8 +77,8 @@ def try_payment_action(order, user, payment, func, *args, **kwargs):
         func(*args, **kwargs)
     except (PaymentError, ValueError) as e:
         message = str(e)
-        OrderEventManager().payment_failed_event(
-            order=order, user=user, message=message, payment=payment).save()
+        payment_failed_event(
+            order=order, user=user, message=message, payment=payment)
         raise ValidationError({'payment': message})
     return True
 
@@ -194,11 +196,10 @@ class OrderAddNote(BaseMutation):
     def perform_mutation(cls, _root, info, **data):
         order = cls.get_node_or_error(
             info, data.get('id'), only_type=Order)
-        event = OrderEventManager().order_note_added_event(
+        event = order_note_added_event(
             order=order, user=info.context.user,
             message=data.get('input')['message'])
-        event.save()
-        return OrderAddNote(order=order, event=event.last)
+        return OrderAddNote(order=order, event=event)
 
 
 class OrderCancel(BaseMutation):
@@ -275,9 +276,9 @@ class OrderCapture(BaseMutation):
             order, info.context.user, payment,
             gateway_capture, payment, amount)
 
-        OrderEventManager().payment_captured_event(
+        payment_captured_event(
             order=order, user=info.context.user,
-            amount=amount, payment=payment).save()
+            amount=amount, payment=payment)
         return OrderCapture(order=order)
 
 
@@ -302,8 +303,8 @@ class OrderVoid(BaseMutation):
         try_payment_action(
             order, info.context.user, payment, gateway_void, payment)
 
-        OrderEventManager().payment_voided_event(
-            order=order, user=info.context.user, payment=payment).save()
+        payment_voided_event(
+            order=order, user=info.context.user, payment=payment)
         return OrderVoid(order=order)
 
 
@@ -334,7 +335,7 @@ class OrderRefund(BaseMutation):
         try_payment_action(
             order, info.context.user, payment, gateway_refund, payment, amount)
 
-        OrderEventManager().payment_refunded_event(
+        payment_refunded_event(
             order=order, user=info.context.user,
-            amount=amount, payment=payment).save()
+            amount=amount, payment=payment)
         return OrderRefund(order=order)
