+++ saleor/payment/utils.py
@@ -12,10 +12,10 @@ from django.utils.translation import pgettext_lazy
 from ..account.models import Address, User
 from ..checkout.models import Checkout
 from ..core import analytics
-from ..events import OrderEventsEmails
-from ..events.order import OrderEventManager
 from ..order import utils as order_utils
 from ..order.emails import send_payment_confirmation
+from ..order.events import order_manually_marked_as_paid_event, \
+    order_fully_paid_event, email_sent_event, OrderEventsEmails
 from ..order.models import Order
 from . import (
     ChargeStatus, CustomPaymentChoices, GatewayError, OperationType,
@@ -76,10 +76,10 @@ def create_payment_information(
 
 
 def handle_fully_paid_order(order):
-    event = OrderEventManager().order_fully_paid_event(order=order)
+    order_fully_paid_event(order=order)
 
     if order.get_user_current_email():
-        event.email_sent_event(
+        email_sent_event(
             order=order, email_type=OrderEventsEmails.PAYMENT, user=None)
         send_payment_confirmation.delay(order.pk)
 
@@ -91,8 +91,6 @@ def handle_fully_paid_order(order):
     except Exception:
         # Analytics failing should not abort the checkout flow
         logger.exception('Recording order in analytics failed')
-
-    event.save()
 
 
 def require_active_payment(view):
@@ -176,8 +174,7 @@ def mark_order_as_paid(order: Order, request_user: User):
     payment.charge_status = ChargeStatus.FULLY_CHARGED
     payment.captured_amount = order.total.gross.amount
     payment.save(update_fields=['captured_amount', 'charge_status'])
-    OrderEventManager().order_manually_marked_as_paid_event(
-        order=order, user=request_user).save()
+    order_manually_marked_as_paid_event(order=order, user=request_user)
 
 
 def create_transaction(
