+++ tests/test_order_utils.py
@@ -1,7 +1,7 @@
 import pytest
 
-from saleor.events.models import OrderEvent
-from saleor.events import OrderEvents
+from saleor.order.events import OrderEvents
+from saleor.order.models import OrderEvent
 from saleor.order.utils import change_order_line_quantity
